HTTP Basic auth support (#6)
## 📝 Summary

Enables basic auth support for API requests. 

The basic auth password is configurable through API and/or file. If set
via API, the salted hash is stored in the file to persist across
reboots.

Config-file updates:

```toml
[general]
# HTTP Basic Auth
basic_auth_secret_path = "basic-auth-secret.txt" # basic auth is supported if a path is provided
basic_auth_secret_salt = "D;%yL9TS:5PalS/d"      # use a random string for the salt
```

`basic_auth_secret_path` specifies the file to store the salted, hashed
secret in. It's loaded (or created) on startup.
- if the file is not empty, API requests need to include a http basic
auth password that matches that sha256 hash (user `admin`)
- if empty, no authentication is required for API requests until secret
is configured through API or file. if `/api/v1/set-basic-auth` is
called, it uses the payload as secret (immediately) and writes the hash
of the secret it to the file (for reuse across restarts).
- if file does not exist, it is created (empty)

Only the salted SHA256 hash of the password is stored, both in the file
as well as in memory.

The secret can be overwritten (updated) via API call, if the request
provides the previous http basic auth secret.

Also added tests and updated the README.

---

## ✅ I have run these commands

* [x] `make lint`
* [x] `make test`
* [x] `go mod tidy`